pyproject.toml
[tool.poetry]
name = "crescendo-ai"
version = "0.1.0"
description = "Raspberry Pi music player that activates based on presence detection"
authors = ["User"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
pyserial = "^3.5"
pyusb = "^1.3.1"
pygame = "^2.6.1"
rpi-gpio = {version = "^0.7.1", markers = "platform_system == 'Linux' and platform_machine.startswith('arm')"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
